finish code splitting
