+++ src/$public/$public.ts
@@ -1,13 +1,19 @@
+import { card } from './card';
 import { config } from './config';
+import { deviceFingerprint } from './fingerprint';
 import { errors } from './errors';
 import { http } from './http';
+import { tokenize } from './tokenize';
 import { utils } from './utils';
 import { validator } from './validator';
 
 export const $public = {
+  card,
   config,
+  deviceFingerprint,
   errors,
   http,
+  tokenize,
   utils,
   validator,
 };
